+++ application/controllers/cash_out.php
@@ -22,6 +22,8 @@ class Cash_out extends CI_Controller {
 		$this->load->library('form_validation');
 		$this->load->model('check_writer_model');
 		$this->load->model('bank_accounts_model');
+		$this->load->model('check_writer_items_model');
+		$this->load->model('syslog_model');
 	}
 	function nomor_bukti($add=false)
 	{
@@ -53,7 +55,8 @@ class Cash_out extends CI_Controller {
 	}
 	function index()
 	{	
-            $this->browse();
+		if(!allow_mod2('_60030'))return false;   
+        $this->browse();
 	}
 	function get_posts(){
             $data=data_table_post($this->table_name);
@@ -61,6 +64,7 @@ class Cash_out extends CI_Controller {
 	}
 	function add()
 	{
+		if(!allow_mod2('_60031'))return false;   
 		 $data=$this->set_defaults();
 		 $this->_set_rules();
  		 $data['mode']='add';
@@ -75,6 +79,8 @@ class Cash_out extends CI_Controller {
 		$this->nomor_bukti(true);
 		$message=mysql_error();
 		if($message=="")$message='Update Success';
+		$this->syslog_model->add($id,"cash_out","add");			
+
 		$this->view($data['voucher'],$message);
 //        header('location: '.base_url().'index.php/cash_out/view/'.$data['voucher']);
 	}
@@ -92,6 +98,8 @@ class Cash_out extends CI_Controller {
             $this->check_writer_model->update($id,$data);
 			$message=mysql_error();
 			if($message=="")$message='Update Success';
+			$this->syslog_model->add($id,"cash_out","edit");			
+
 		} else {
 			$message=mysql_error();
 			if($message=="")$message='Update Gagal';
@@ -100,6 +108,7 @@ class Cash_out extends CI_Controller {
 //        header('location: '.base_url().'index.php/cash_out/view/'.$data['voucher']);
 	}
 	function view($id,$message=null){
+		if(!allow_mod2('_60030'))return false;   
 		$id=urldecode($id);
 		 $data['id']=$id;
 		 $model=$this->check_writer_model->get_by_id($id)->row();
@@ -168,7 +177,7 @@ class Cash_out extends CI_Controller {
 				}
 			}
 		}
-        $sql.=" limit $offset,$limit";
+        //$sql.=" limit $offset,$limit";
         echo datasource($sql);
     }	 
 	function items($voucher) {
@@ -187,6 +196,9 @@ class Cash_out extends CI_Controller {
 			echo json_encode(array('success'=>false,'msg'=>'Nomor voucher tidak ada atau kosong.'));
 			return false;
 		}
+		if(!$line_number=$this->input->post("line_number")){
+			$line_number=0;
+		}
 		$this->load->model('check_writer_model');
 		$trans_id=$this->check_writer_model->get_by_id($voucher)->row()->trans_id;
 		
@@ -200,28 +212,54 @@ class Cash_out extends CI_Controller {
 		$data['description']=$coa->account_description;
 		$data['amount']=$this->input->post('amount');
 		$data['comments']=$this->input->post('comments');
-		if($this->check_writer_items_model->save($data)){
-			echo json_encode(array('success'=>true,'msg'=>'Sukses tambah data.'));
-			return true;
+		
+		$ret['success']=false;
+		$ret['msg']="Unknown Error";
+		if($line_number==0){
+			if($ok = $this->check_writer_items_model->save($data)){
+				$ret['success']=true;
+				$ret['msg']='Sukses tambah data';
+			}
 		} else {
-			echo json_encode(array('success'=>false,'msg'=>'Gagal menyimpan data.'));
-			return false;
+			if($ok = $this->check_writer_items_model->update($line_number,$data)){
+				$ret['success']=true;
+				$ret['msg']='Sukses tambah data';
+			}			
 		}
+		echo json_encode($ret);
+		return $ok;
+	}
+	function delete_item(){
+		$id=$this->input->post("line_number");
+		$data["success"]=false;
+		$data["msg"]="Unknown Error !";
+		if($this->check_writer_items_model->delete($id)){
+			$data["success"]=true;
+			$data["msg"]="Berhasil.";
+			$this->syslog_model->add($id,"cash_out","delete");
+
+		}
+		echo json_encode($data);
 	}
 	function unposting($voucher) {
+		if(!allow_mod2('_60035'))return false;   
 		$voucher=urldecode($voucher);
 		$message=$this->check_writer_model->unposting($voucher);
 		$this->view($voucher,$message);
 	}
 	function posting($voucher) {
+		if(!allow_mod2('_60035'))return false;   
 		$voucher=urldecode($voucher);
 		$message=$this->check_writer_model->posting($voucher);
 		$this->view($voucher,$message);
 	}
 	function delete($voucher) {
+		if(!allow_mod2('_60033'))return false;   
 		$voucher=urldecode($voucher);
 		$message=$this->check_writer_model->delete($voucher);
 		if($message!=""){
+			$this->syslog_model->add($voucher,"cash_out","delete");
+
 			$this->view($voucher,$message);
 			return false;
 		}
